requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.3.0
seaborn>=0.11.0
jupyter>=1.0.0
textblob>=0.17.0
scikit-learn>=1.0.0
openai>=1.40.0
python-dotenv>=1.0.0
Flask>=3.0.0
